MOD: modularize euler program.
