+++ chapter2/EULER/input.txt
@@ -1,9 +1,12 @@
-4 6 
-2
+5 10
+4
 0 1
 0 2
 0 3
+0 4
 1 2
 1 3
+1 4
 2 3
+2 4
 3 4
